base layer: refactor folder structure
