+++ frontend/src/features/map_planning/routes/Map.tsx
@@ -4,7 +4,7 @@ import { BaseStage } from '../components/BaseStage';
 import { Layers } from '../components/toolbar/Layers';
 import { PlantSearch } from '../components/toolbar/PlantSearch';
 import { Toolbar } from '../components/toolbar/Toolbar';
-import BaseLayer from '../layers/BaseLayer';
+import BaseLayer from '../layers/base/BaseLayer';
 import PlantsLayer from '../layers/plant/PlantsLayer';
 import useMapStore from '../store/MapStore';
 import { LayerName } from '../store/MapStoreTypes';
@@ -15,7 +15,7 @@ import SimpleButton from '@/components/Button/SimpleButton';
 import SimpleFormInput from '@/components/Form/SimpleFormInput';
 import { baseApiUrl } from '@/config';
 import { createNextcloudWebDavClient } from '@/config/nextcloud_client';
-import { UpdateBaseLayerAction } from '@/features/map_planning/layers/plant/actions';
+import { UpdateBaseLayerAction } from '@/features/map_planning/layers/base/actions';
 import { useSafeAuth } from '@/hooks/useSafeAuth';
 import { ReactComponent as ArrowIcon } from '@/icons/arrow.svg';
 import { ReactComponent as MoveIcon } from '@/icons/move.svg';
